Remove most of the notebook tool conversion/ diff logic from the participant in favor of helper functions
